requirements.txt
anthropic
requests
pymupdf
pdfminer.six
pymupdf4llm
openpyxl
python-dotenv
msal
pyyaml
